#49 Implement iteration with HANDOUT_SFIX controlling behavior
In this iteration, we operate on the implicit assumption that the target[0] is the path at which the clean pdf is to be created. A construction string variable called HANDOUT_SFIX controls whether or not the handout pdf is generated. When HANDOUT_SFIX is set to empty (default), only the clean pdf is generated (no handout). When HANDOUT_SFIX is not empty, we check whether any paths in the target list have a tail that matches {the tail of target[0] with HANDOUT_SFIX inserted immediately before the file extension}. If check returns true, we generate handout pdfs at matching paths (as well as clean pdf at target[0]). If check returns false, we raise an exception.