#49 Implement iteration with target list length controlling behavior
This iteration uses a combination of the target list length and HANDOUT_SFIX in determining whether and where to create a handout. Compared to the previous version of this code, there is weaker restriction on the handout file name: the only requirement is that if HANDOUT_SFIX is nonempty, the file name must end with "[HANDOUT_SFIX].pdf". Refer to issue comment #1192929420 for more details on case-by-case behavior.