PR for #49: Handout functionality for LyX builder (#62)
* #49 Creating template slides

* #49 Produces handout from talk_slide.pdf

* #49 Only builds handout if in targets

* #49 Cleaning implementation of handout maker

* #49 Adding cleanup function for produced handout lyx

* #49 Full run of repo with handout edits

* #49 Storing handout in /temp/

* #49 Working on using argument for behavior control

* #49 Add variable to control handout behavior

* #49 Change control variable to boolean

* #49 Delete unused function

* #49 Change control variable type to T/F

* #49 Fix error when creating /paper/ handout

* #49 Modify create_handout() structure

* #49 Implement iteration with HANDOUT_SFIX controlling behavior

In this iteration, we operate on the implicit assumption that the target[0] is the path at which the clean pdf is to be created. A construction string variable called HANDOUT_SFIX controls whether or not the handout pdf is generated. When HANDOUT_SFIX is set to empty (default), only the clean pdf is generated (no handout). When HANDOUT_SFIX is not empty, we check whether any paths in the target list have a tail that matches {the tail of target[0] with HANDOUT_SFIX inserted immediately before the file extension}. If check returns true, we generate handout pdfs at matching paths (as well as clean pdf at target[0]). If check returns false, we raise an exception.

* #49 Implement iteration with target list length controlling behavior

This iteration uses a combination of the target list length and HANDOUT_SFIX in determining whether and where to create a handout. Compared to the previous version of this code, there is weaker restriction on the handout file name: the only requirement is that if HANDOUT_SFIX is nonempty, the file name must end with "[HANDOUT_SFIX].pdf". Refer to issue comment #1192929420 for more details on case-by-case behavior.

* #49 Update tests to include handout functionality

* #49 Refactor do_call() function

* #49 Make path matching case insensitive

* #49 Change handout options

Co-authored-by: jmshapir <jesse.m.shapiro@gmail.com>

* #49 Add blank line at end of file

* #49 Rename talk_slides.lyx to TemplateSlides.lyx

* #49 Add PDF output of slides to .gitignore

* #49 Remove /temp/ file from .gitignore

* #49 Add blank line at end of file

* #49 Move construction variable definition outside of Sconstruct

* #49 Add test for undeclared HANDOUT_SFIX

* #49 Move handout_standard_test() to test_build_lyx

* #49 Improve error message

Co-authored-by: Richard Calvo <97416734+rcalvo12@users.noreply.github.com>

* #49 Improve error message

Co-authored-by: Richard Calvo <97416734+rcalvo12@users.noreply.github.com>

* #49 Improve error message

Co-authored-by: Richard Calvo <97416734+rcalvo12@users.noreply.github.com>

Co-authored-by: Richard Calvo <97416734+rcalvo12@users.noreply.github.com>
Co-authored-by: jmshapir <jesse.m.shapiro@gmail.com>